rename pubsub-update to file-status-update
